+++ NEWS.md
@@ -1,6 +1,7 @@
 # lstcompr 0.1.1
 
 * variable ranges can also be of character type
+* fixed sorting of free variables
 
 # lstcompr 0.1.0